Sprucing up a few chapters (A-6, B-3, 3, and D-3)
Hub A: Added a 'dark' tile that's the transparent one, using it in the tavern so that we aren't using a weird one as the 'background'.
A-6: Modified the ground of the cave to be more varied, adjusted shading, added transition tiles for wood/cave.
B-3: Added a stone trim to the main fort to help spruce up the map, including in cutscenes (and slightly adjusted geometry to accommodate)
Ch 3: Added cosmetic 'terminals' that are what control the platforms (though the player can't interact with them), as well as refining the platforms with adjusting planks/adding columns.
D3: General spruce-up, more detailed ground in the main village area (no adjustment to map flow though)